Migrate seeds from size_t to uint32_t (#4965)

+++ src/engine/rand.h
@@ -126,14 +126,14 @@ namespace Rand
     class DeterministicRandomGenerator
     {
     public:
-        explicit DeterministicRandomGenerator( const size_t initialSeed );
+        explicit DeterministicRandomGenerator( const uint32_t initialSeed );
 
         // prevent accidental copies
         DeterministicRandomGenerator( const DeterministicRandomGenerator & ) = delete;
         DeterministicRandomGenerator & operator=( const DeterministicRandomGenerator & ) = delete;
 
-        size_t GetSeed() const;
-        void UpdateSeed( const size_t seed );
+        uint32_t GetSeed() const;
+        void UpdateSeed( const uint32_t seed );
 
         uint32_t Get( const uint32_t from, const uint32_t to = 0 ) const;
 
@@ -141,7 +141,7 @@ namespace Rand
         const T & Get( const std::vector<T> & vec ) const
         {
             ++_currentSeed;
-            std::mt19937 seededGen( static_cast<uint32_t>( _currentSeed ) );
+            std::mt19937 seededGen( _currentSeed );
             return Rand::GetWithGen( vec, seededGen );
         }
 
@@ -149,11 +149,11 @@ namespace Rand
         void Shuffle( std::vector<T> & vector ) const
         {
             ++_currentSeed;
-            Rand::ShuffleWithSeed( vector, static_cast<uint32_t>( _currentSeed ) );
+            Rand::ShuffleWithSeed( vector, _currentSeed );
         }
 
     private:
-        mutable size_t _currentSeed; // this is mutable so clients that only call RNG method can receive a const instance
+        mutable uint32_t _currentSeed; // this is mutable so clients that only call RNG method can receive a const instance
     };
 }
